# Optimize PIDSet performance